Simply terminate the connection if too many requests were issued
The previous solution erroneously generates an error message if the
request limit is exceeded. This solution simply terminates the
connection rather than tell the client that there was an error on its
end. A client may re-establish a connection and issue further
requests, but the termination of the request handler is secured.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>