Close connection if HTTP version < 1.1 or explicity requested
Turns out some older HTTP/1.0 clients wait for the server to send
more data over if the server keeps the connection open (irregardless
of the Content-Length field). Those clients will always wait until
a SIGALRM timeout is triggered on the server side, causing the user
in front of the client to complain about an insanely slow server.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>